Create a default PR template for the organization (#4)
* Create a default PR template for the organization

It is useful to provide guidance and structure for someone making pull request (PR), making it easier for them to know what information is expected. This PR adds a template for the PR description to help provide structure for contributors. Since this is being created the organizational repository (RxnRover/.github), it will be applied to all repos in the organization, unless they override it with their own template.

* Update wording and add a section for related issues and PRs